pyproject.toml
[tool.poetry]
name = "runtime-settings"
version = "0.0.1"
description = "Server for storing service settings."
license = "MIT"
authors = ["Alexey Petrunnik <petrunnik.a@gmail.com>"]
packages = [{include = "runtime_config", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.1"
uvicorn = { version = "^0.18.3", extras = ["standard"] }
pydantic = { version = "^1.10.2", extras = ["dotenv"] }
aiopg = "^1.4.0"
psycopg2 = "^2.9.5"
types-psycopg2 = "^2.9.21"
SQLAlchemy = { extras = ["mypy"], version = "^1.4.45" }
alembic = "^1.9.1"
structlog = "^22.3.0"
passlib = {version = "^1.7.4", extras = ["argon2"]}
types-passlib = "^1.7.7.3"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
types-python-jose = "^3.3.4"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.18.3"
pytest-mock = "^3.10.0"
pytest-cov = "^3.0.0"
autopep8 = "1.6.0"
black = "22.8.0"
flake8 = "4.0.1"
isort = "5.10.1"
mypy = "0.971"
pre-commit = "^2.21.0"
httpx = "^0.23.1"

[tool.poetry.scripts]
runtime-config = "runtime_config.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ['./src']
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "src/runtime_config/cli.py"
]

[tool.black]
target_version = ['py310']
line-length = 119
skip_string_normalization = true

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
disallow_any_generics = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module="uvicorn.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
